refactor(subscription): Split mod.rs into focused sub-modules
Split the 2,692 line subscription/mod.rs into smaller, focused modules:
- config.rs (164 lines): SubscriptionConfig, SubscriptionRetryPolicy
- types.rs (456 lines): SubscriptionId, Subscription, SubscriptionMetrics, SubscriptionUpdate, SubscriptionError
- delta.rs (286 lines): PartialRowDelta, compute_delta, compute_delta_with_pk
- selective.rs (346 lines): SelectiveColumnConfig, ColumnDiff, selective update functions
- hash.rs (42 lines): hash_rows, hash_row

The refactored mod.rs (1,489 lines) now serves as a thin re-export layer
plus integration tests. All 176 subscription tests pass, and all public
APIs are maintained via pub use re-exports.

Closes #4122
